refactor: ApplicationContext 클래스 이름 변경
